+++ src/dotnet-inspect/Commands/ProjectCommand.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
-using System.Text;
 using DotnetInspector.Core;
+using DotnetInspector.Inspectors;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
 using DotnetInspector.Packages;
@@ -15,16 +15,7 @@ namespace DotnetInspector.Commands;
 
 public class ProjectCommand
 {
-    enum ProjectSkillReadFailure
-    {
-        None,
-        InvalidName,
-        InvalidDescription,
-    }
-
     public const string Name = "project";
-    static readonly string[] ProjectReadmeCandidates = ["README.md", "PROJECT.md"];
-    static readonly string[] SkillPatterns = ["skills/SKILL.md", "skills/**/SKILL.md"];
 
     public static async Task<int> ExecuteAsync(
         ProjectOptions options,
@@ -170,23 +161,26 @@ public class ProjectCommand
             schema,
             candidateSections,
             selectedValueField);
-        ProjectQueryContext queryContext = new(
-            () => ReadSkills(
-                assetsPath,
-                options.Tfm,
-                deferDocumentContent,
-                contentStore),
-            () => ReadAgentGuidance(
+        var skillsProvider = new ProjectSkillsProvider(
+            assetsPath,
+            options.Tfm,
+            deferDocumentContent,
+            contentStore);
+        var agentGuidanceProvider =
+            new ProjectAgentGuidanceProvider(
                 dependencies,
                 deferDocumentContent,
-                contentStore),
-            token => ReadPackageDocumentsAsync(
-                focusedDependencies,
-                options,
-                commandContext,
-                deferContent: true,
-                contentStore,
-                token));
+                contentStore);
+        var packageDocumentsProvider = new ProjectPackageDocumentsProvider(
+            focusedDependencies,
+            options.SourceOptions,
+            commandContext,
+            deferDocumentContent,
+            contentStore);
+        ProjectQueryContext queryContext = new(
+            skillsProvider.Read,
+            agentGuidanceProvider.Read,
+            packageDocumentsProvider.ReadAsync);
         InspectionQueryResults queryResults = await catalog.QueryRegistry.RunAsync(
             requestedQueries,
             queryContext,
@@ -730,498 +724,9 @@ public class ProjectCommand
     {
         foreach (ProjectContentFailure failure in failures)
         {
-            if (failure.RedactIdentity)
-                CommandError.Write(failure.Reason);
-            else
-                CommandError.WriteWarning(
-                    $"Could not read '{failure.Package}' file '{failure.Path}': {failure.Reason}");
+            CommandError.WriteWarning(
+                $"Could not read '{failure.Package}' file '{failure.Path}': {failure.Reason}");
         }
-    }
-
-    static ProjectSkillsResult ReadSkills(
-        string assetsPath,
-        string? tfm,
-        bool deferContent,
-        ProjectDocumentContentStore contentStore)
-    {
-        var skills = new List<ProjectSkillData>();
-        var failures = new List<ProjectContentFailure>();
-        foreach (ProjectPackageFileEntry file in
-                 ProjectAssetsParser.ParsePackageFileEntries(
-                     assetsPath,
-                     tfm,
-                     SkillPatterns,
-                     log: null))
-        {
-            if (string.IsNullOrWhiteSpace(file.FullPath)
-                || !File.Exists(file.FullPath))
-            {
-                failures.Add(new ProjectContentFailure(
-                    file.PackageName,
-                    file.Path,
-                    "a restored package skill listed in project.assets.json is missing from the package cache"));
-                skills.Add(new ProjectSkillData(
-                    file.PackageName,
-                    file.Version,
-                    file.Path,
-                    null,
-                    "",
-                    "",
-                    null));
-                continue;
-            }
-
-            long? size = null;
-            try
-            {
-                long knownSize = new FileInfo(file.FullPath).Length;
-                size = knownSize;
-                string content = File.ReadAllText(file.FullPath);
-                IReadOnlyDictionary<string, string> frontmatter =
-                    MarkdownContent.ParseYamlFrontmatter(content);
-                ProjectSkillReadFailure validationFailure =
-                    ValidateSkillMetadata(
-                        file.Path,
-                        frontmatter,
-                        out string skillName,
-                        out string description);
-                if (validationFailure != ProjectSkillReadFailure.None)
-                {
-                    failures.Add(new ProjectContentFailure(
-                        file.PackageName,
-                        file.Path,
-                        validationFailure == ProjectSkillReadFailure.InvalidName
-                            ? "a restored package skill must declare an Agent Skills-compliant name that matches its containing directory"
-                            : "a restored package skill must declare an Agent Skills-compliant description of 1 to 1024 characters",
-                        RedactIdentity: true));
-                    continue;
-                }
-
-                if (deferContent)
-                {
-                    contentStore.Add(
-                        ProjectSectionNames.Skills,
-                        file.PackageName,
-                        file.Path,
-                        file.FullPath);
-                    skills.Add(new ProjectSkillData(
-                        file.PackageName,
-                        file.Version,
-                        file.Path,
-                        knownSize,
-                        skillName,
-                        description,
-                        ""));
-                    continue;
-                }
-
-                contentStore.Add(
-                    ProjectSectionNames.Skills,
-                    file.PackageName,
-                    file.Path,
-                    file.FullPath);
-                skills.Add(new ProjectSkillData(
-                    file.PackageName,
-                    file.Version,
-                    file.Path,
-                    knownSize,
-                    skillName,
-                    description,
-                    ""));
-            }
-            catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-            {
-                failures.Add(new ProjectContentFailure(
-                    file.PackageName,
-                    file.Path,
-                    ex.Message));
-                if (size is long knownSize)
-                {
-                    skills.Add(new ProjectSkillData(
-                        file.PackageName,
-                        file.Version,
-                        file.Path,
-                        knownSize,
-                        "",
-                        "",
-                        null));
-                }
-            }
-        }
-
-        return ProjectSkillsQuery.Execute(skills, failures);
-    }
-
-    static ProjectSkillReadFailure ValidateSkillMetadata(
-        string packagePath,
-        IReadOnlyDictionary<string, string> frontmatter,
-        out string name,
-        out string description)
-    {
-        if (!TryGetSkillName(packagePath, frontmatter, out name))
-        {
-            description = "";
-            return ProjectSkillReadFailure.InvalidName;
-        }
-
-        frontmatter.TryGetValue("description", out string? candidateDescription);
-        if (!IsAgentSkillDescription(candidateDescription))
-        {
-            description = "";
-            return ProjectSkillReadFailure.InvalidDescription;
-        }
-
-        description = candidateDescription!;
-        return ProjectSkillReadFailure.None;
-    }
-
-    static bool TryGetSkillName(
-        string packagePath,
-        IReadOnlyDictionary<string, string> frontmatter,
-        out string name)
-    {
-        string normalizedPath = packagePath.Replace('\\', '/');
-        int fileSeparator = normalizedPath.LastIndexOf('/');
-        if (fileSeparator <= 0)
-        {
-            name = "";
-            return false;
-        }
-
-        string parentPath = normalizedPath[..fileSeparator].TrimEnd('/');
-        int parentSeparator = parentPath.LastIndexOf('/');
-        string directoryName = parentPath[(parentSeparator + 1)..];
-        if (!frontmatter.TryGetValue("name", out name!))
-            return false;
-
-        return string.Equals(name, directoryName, StringComparison.Ordinal)
-            && IsAgentSkillName(name);
-    }
-
-    static bool IsAgentSkillName(string name)
-    {
-        if (name.Length is < 1 or > 64
-            || name[0] == '-'
-            || name[^1] == '-')
-        {
-            return false;
-        }
-
-        bool previousWasHyphen = false;
-        foreach (char character in name)
-        {
-            bool isHyphen = character == '-';
-            if (!(character is >= 'a' and <= 'z'
-                  || character is >= '0' and <= '9'
-                  || isHyphen)
-                || isHyphen && previousWasHyphen)
-            {
-                return false;
-            }
-
-            previousWasHyphen = isHyphen;
-        }
-
-        return true;
-    }
-
-    static bool IsAgentSkillDescription(string? description)
-    {
-        if (string.IsNullOrEmpty(description))
-            return false;
-
-        int length = 0;
-        foreach (Rune _ in description.EnumerateRunes())
-        {
-            if (++length > 1024)
-                return false;
-        }
-
-        return true;
-    }
-
-    static ProjectAgentGuidanceResult ReadAgentGuidance(
-        IReadOnlyList<ProjectPackageReference> dependencies,
-        bool deferContent,
-        ProjectDocumentContentStore contentStore)
-    {
-        var guidance = new List<ProjectAgentGuidanceData>();
-        var failures = new List<ProjectContentFailure>();
-        foreach (ProjectPackageReference dependency in dependencies)
-        {
-            const string relativePath = "AGENTS.md";
-            string? fullPath = string.IsNullOrWhiteSpace(dependency.PackagePath)
-                ? null
-                : Path.Combine(dependency.PackagePath, relativePath);
-            if (fullPath is null || !File.Exists(fullPath))
-            {
-                guidance.Add(new ProjectAgentGuidanceData(
-                    dependency.PackageName,
-                    dependency.Version,
-                    "",
-                    "",
-                    "",
-                    null));
-                continue;
-            }
-
-            try
-            {
-                if (deferContent)
-                {
-                    contentStore.Add(
-                        ProjectSectionNames.AgentGuidance,
-                        dependency.PackageName,
-                        relativePath,
-                        fullPath);
-                    guidance.Add(new ProjectAgentGuidanceData(
-                        dependency.PackageName,
-                        dependency.Version,
-                        relativePath,
-                        "",
-                        "",
-                        ""));
-                    continue;
-                }
-
-                string content = File.ReadAllText(fullPath);
-                IReadOnlyDictionary<string, string> frontmatter =
-                    MarkdownContent.ParseYamlFrontmatter(content);
-                frontmatter.TryGetValue("name", out string? name);
-                frontmatter.TryGetValue("description", out string? description);
-                contentStore.Add(
-                    ProjectSectionNames.AgentGuidance,
-                    dependency.PackageName,
-                    relativePath,
-                    fullPath);
-                guidance.Add(new ProjectAgentGuidanceData(
-                    dependency.PackageName,
-                    dependency.Version,
-                    relativePath,
-                    name ?? "",
-                    description ?? "",
-                    ""));
-            }
-            catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-            {
-                failures.Add(new ProjectContentFailure(
-                    dependency.PackageName,
-                    relativePath,
-                    ex.Message));
-                guidance.Add(new ProjectAgentGuidanceData(
-                    dependency.PackageName,
-                    dependency.Version,
-                    relativePath,
-                    "",
-                    "",
-                    ""));
-            }
-        }
-
-        return ProjectAgentGuidanceQuery.Execute(guidance, failures);
-    }
-
-    static async ValueTask<ProjectPackageDocumentsResult> ReadPackageDocumentsAsync(
-        IReadOnlyList<ProjectPackageReference> dependencies,
-        ProjectOptions options,
-        CommandContext context,
-        bool deferContent,
-        ProjectDocumentContentStore contentStore,
-        CancellationToken cancellationToken)
-    {
-        var documents = new List<ProjectPackageDocumentData>();
-        var failures = new List<ProjectContentFailure>();
-        foreach (ProjectPackageReference dependency in dependencies)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            var acquired = await ReadBestPackageDocumentAsync(
-                dependency,
-                options.SourceOptions,
-                context,
-                deferContent,
-                contentStore,
-                cancellationToken);
-            if (acquired.Document is not null)
-                documents.Add(acquired.Document);
-            if (acquired.Failure is not null)
-                failures.Add(acquired.Failure);
-        }
-
-        return ProjectPackageDocumentsQuery.Execute(documents, failures);
-    }
-
-    static async Task<(
-        ProjectPackageDocumentData? Document,
-        ProjectContentFailure? Failure)> ReadBestPackageDocumentAsync(
-        ProjectPackageReference dependency,
-        NuGetSourceOptions? sourceOptions,
-        CommandContext context,
-        bool deferContent,
-        ProjectDocumentContentStore contentStore,
-        CancellationToken cancellationToken)
-    {
-        if (!string.IsNullOrWhiteSpace(dependency.PackagePath)
-            && Directory.Exists(dependency.PackagePath))
-        {
-            try
-            {
-                return (
-                    ReadBestPackageDocumentFromDirectory(
-                        dependency,
-                        deferContent,
-                        contentStore),
-                    null);
-            }
-            catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-            {
-                return (
-                    null,
-                    new ProjectContentFailure(
-                        dependency.PackageName,
-                        "README.md|PROJECT.md",
-                        ex.Message));
-            }
-        }
-
-        PackageExtractionResult? resolution = null;
-        bool retainTemporaryDirectory = false;
-        try
-        {
-            // Package acquisition is single-flight. Cancel this caller's wait
-            // without cancelling the shared acquisition for other callers.
-            PackageExtractionOutcome outcome =
-                await PackageExtractor.ExtractPackageAsync(
-                        context.HttpClient,
-                        dependency.PackageName,
-                        context.Logger.Log,
-                        sourceOptions: sourceOptions,
-                        version: dependency.Version)
-                    .WaitAsync(cancellationToken);
-            cancellationToken.ThrowIfCancellationRequested();
-            if (!outcome.IsSuccess)
-            {
-                return (
-                    null,
-                    new ProjectContentFailure(
-                        dependency.PackageName,
-                        "README.md|PROJECT.md",
-                        outcome.ErrorMessage ?? "package acquisition failed"));
-            }
-
-            resolution = outcome.Result!;
-            var resolved = dependency with
-            {
-                Version = resolution.Version ?? dependency.Version,
-                PackagePath = resolution.ExtractPath,
-            };
-            string? cleanupDirectory = resolution is
-            {
-                FromCache: false,
-                TempDir: not null,
-            }
-                ? resolution.TempDir
-                : null;
-            ProjectPackageDocumentData? document =
-                ReadBestPackageDocumentFromDirectory(
-                    resolved,
-                    deferContent,
-                    contentStore,
-                    cleanupDirectory);
-            retainTemporaryDirectory =
-                deferContent
-                && document is not null
-                && cleanupDirectory is not null;
-            return (document, null);
-        }
-        catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-        {
-            return (
-                null,
-                new ProjectContentFailure(
-                    dependency.PackageName,
-                    "README.md|PROJECT.md",
-                    ex.Message));
-        }
-        finally
-        {
-            if (!retainTemporaryDirectory
-                && resolution is { FromCache: false, TempDir: not null }
-                && Directory.Exists(resolution.TempDir))
-            {
-                try
-                {
-                    Directory.Delete(resolution.TempDir, recursive: true);
-                }
-                catch (Exception ex) when (ex is IOException or UnauthorizedAccessException)
-                {
-                    context.Logger.LogWarning(
-                        $"Could not remove temporary package directory "
-                        + $"'{resolution.TempDir}': {ex.Message}");
-                }
-            }
-        }
-    }
-
-    static ProjectPackageDocumentData? ReadBestPackageDocumentFromDirectory(
-        ProjectPackageReference dependency,
-        bool deferContent,
-        ProjectDocumentContentStore contentStore,
-        string? cleanupDirectory = null)
-    {
-        if (string.IsNullOrWhiteSpace(dependency.PackagePath)
-            || !Directory.Exists(dependency.PackagePath))
-        {
-            return null;
-        }
-
-        string? readme = ResolveProjectReadme(dependency.PackagePath);
-        if (readme is null)
-            return null;
-
-        string fullPath = Path.Combine(
-            dependency.PackagePath,
-            readme.Replace('/', Path.DirectorySeparatorChar));
-        if (!File.Exists(fullPath))
-            return null;
-
-        long size = new FileInfo(fullPath).Length;
-        if (deferContent)
-        {
-            contentStore.Add(
-                ProjectSectionNames.PackageDocs,
-                dependency.PackageName,
-                readme,
-                fullPath,
-                cleanupDirectory);
-        }
-
-        return new ProjectPackageDocumentData(
-            dependency.PackageName,
-            dependency.Version,
-            readme,
-            size,
-            deferContent ? "" : File.ReadAllText(fullPath));
-    }
-
-    static string? ResolveProjectReadme(string packagePath)
-    {
-        foreach (string candidate in ProjectReadmeCandidates)
-        {
-            string? match = Directory.EnumerateFiles(
-                    packagePath,
-                    "*",
-                    SearchOption.TopDirectoryOnly)
-                .FirstOrDefault(file => Path.GetFileName(file).Equals(
-                    candidate,
-                    StringComparison.OrdinalIgnoreCase));
-            if (match is not null)
-            {
-                return Path.GetRelativePath(packagePath, match)
-                    .Replace('\\', '/');
-            }
-        }
-
-        return null;
     }
 
     static int PrintDocument(
@@ -1546,7 +1051,7 @@ public class ProjectCommand
             "package" => row.Package,
             "version" => row.Version,
             "path" => row.Path,
-            "size" => row.Size?.ToString(CultureInfo.InvariantCulture),
+            "size" => row.Size.ToString(CultureInfo.InvariantCulture),
             "name" => row.Name,
             "description" => row.Description,
             _ => row.Path,
@@ -1640,73 +1145,4 @@ public class ProjectCommand
             Console.Write(output);
     }
 
-    sealed class ProjectDocumentContentStore(CommandContext context) : IDisposable
-    {
-        readonly Dictionary<DocumentKey, string> _paths = [];
-        readonly HashSet<string> _cleanupDirectories =
-            new(StringComparer.Ordinal);
-
-        public void Add(
-            string section,
-            string package,
-            string path,
-            string fullPath,
-            string? cleanupDirectory = null)
-        {
-            _paths[new DocumentKey(section, package, path)] = fullPath;
-            if (cleanupDirectory is not null)
-                _cleanupDirectories.Add(cleanupDirectory);
-        }
-
-        public bool Contains(string section, string package, string path)
-            => _paths.ContainsKey(new DocumentKey(section, package, path));
-
-        public string? Read(string section, string package, string path)
-        {
-            if (!_paths.TryGetValue(
-                    new DocumentKey(section, package, path),
-                    out string? fullPath))
-            {
-                return null;
-            }
-
-            try
-            {
-                return File.ReadAllText(fullPath);
-            }
-            catch (Exception ex) when (
-                ex is IOException or UnauthorizedAccessException)
-            {
-                CommandError.WriteWarning(
-                    $"Could not read '{package}' file '{path}': {ex.Message}");
-                return null;
-            }
-        }
-
-        public void Dispose()
-        {
-            foreach (string directory in _cleanupDirectories)
-            {
-                if (!Directory.Exists(directory))
-                    continue;
-
-                try
-                {
-                    Directory.Delete(directory, recursive: true);
-                }
-                catch (Exception ex) when (
-                    ex is IOException or UnauthorizedAccessException)
-                {
-                    context.Logger.LogWarning(
-                        $"Could not remove temporary package directory "
-                        + $"'{directory}': {ex.Message}");
-                }
-            }
-        }
-
-        readonly record struct DocumentKey(
-            string Section,
-            string Package,
-            string Path);
-    }
 }
